extract firebase initialization
