+++ src/modules/Login.js
@@ -1,6 +1,6 @@
 import React from "react";
 import firebase from "firebase/app";
-// import "firebase/firestore";
+import { withRouter } from 'react-router-dom';
 
 class Login extends React.Component {
   constructor(props) {
@@ -65,4 +65,4 @@ class Login extends React.Component {
   }
 }
 
-export default Login;
+export default withRouter(Login);
